analysis/experiment_executor/utils.py
from pathlib import Path
from typing import Union

import yaml


def load_yaml(path: Union[Path, str]) -> dict:
    path = Path(path)
    with path.open("r") as f:
        return yaml.load(f, Loader=yaml.FullLoader)